Add worker to fetch known accounts and assets
