+++ src/workers/fetch-worker.ts
@@ -1,0 +1,65 @@
+import { expose } from "threads"
+import { AccountRecord } from "../lib/stellar-expert"
+import { AssetRecord } from "../lib/stellar-ticker"
+
+const fetcher = {
+  fetchWellknownAccounts,
+  fetchAllAssets
+}
+
+export type Fetcher = typeof fetcher
+
+expose(fetcher)
+
+async function fetchWellknownAccounts(testnet: boolean): Promise<AccountRecord[]> {
+  const requestURL = testnet
+    ? "https://api.stellar.expert/api/explorer/testnet/directory"
+    : "https://api.stellar.expert/api/explorer/public/directory"
+
+  const response = await fetch(requestURL)
+
+  if (response.status >= 400) {
+    throw Error(`Bad response (${response.status}) from stellar.expert server`)
+  }
+
+  const json = await response.json()
+  const knownAccounts = json._embedded.records as AccountRecord[]
+  return knownAccounts
+}
+
+function byAccountCountSorter(a: AssetRecord, b: AssetRecord) {
+  return b.num_accounts - a.num_accounts
+}
+
+function trimAccountRecord(record: AssetRecord) {
+  return {
+    code: record.code,
+    desc: record.desc,
+    issuer: record.issuer,
+    issuer_detail: {
+      name: record.issuer_detail.name,
+      url: record.issuer_detail.url
+    },
+    name: record.name,
+    num_accounts: record.num_accounts,
+    status: record.status,
+    type: record.type
+  }
+}
+
+async function fetchAllAssets(testnet: boolean): Promise<AssetRecord[]> {
+  const requestURL = testnet
+    ? "https://ticker-testnet.stellar.org/assets.json"
+    : "https://ticker.stellar.org/assets.json"
+
+  const response = await fetch(requestURL)
+
+  if (response.status >= 400) {
+    throw Error(`Bad response (${response.status}) from stellar.expert server`)
+  }
+
+  const json = await response.json()
+  const allAssets = json.assets as AssetRecord[]
+  const abbreviatedAssets = allAssets.sort(byAccountCountSorter).map(record => trimAccountRecord(record))
+  return abbreviatedAssets
+}
